pyproject.toml
[tool.poetry]
name = "replicate-flux-proxy"
version = "0.1.0"
description = ""
authors = ["sugarforever <william.1022@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
